Documentation for public methods in the protocols
