+++ SearchApp/SearchApp/Constants.swift
@@ -24,4 +24,32 @@ struct Constants {
     static let persistanceKey = "SearchTerms"
     static let suggestionsCell = "suggestionsCell"
     static let searchBarPlaceholder = "Search Images"
+    static let imagePreviewViewController = "ImagePreviewViewController"
+    static let mainStoryboard = "Main"
+}
+
+enum SearchErrors: Error {
+    case noInternet
+    case parsingError
+    case noData
+    case empty
+    
+    public var errorDescription: String {
+        switch self {
+        case .noInternet:
+            return ErrorStrings.noInternetError
+        case .parsingError:
+            return ErrorStrings.somethingWentWrong
+        case .noData:
+            return ErrorStrings.noResultsFound
+        case .empty:
+            return ErrorStrings.startSearching
+        }
+    }
+    
+}
+
+enum ResultType {
+    case success(imageModel: ImageResponseModel)
+    case failure(error:SearchErrors?)
 }
